move all header type and functions to new header.go file
